# api-spec.yaml
openapi: 3.0.3
info:
  title: Saitynai API
  version: 1.0.0
  description: |
    REST API for "Saitynai" community platform.
    Supports user registration, authentication, communities, posts, comments and admin operations (roles, blocking users).
servers:
  - url: https://saitynai-vl9u5.ondigitalocean.app/
    description: Production (DigitalOcean)
  - url: http://localhost:5000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /User/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterDTO"
            example:
              username: "jonas"
              email: "jonas@example.com"
              password: "Jonas123!"
              name: "Jonas"
              surname: "Jonaitis"
      responses:
        "201":
          description: User created successfully
        "400":
          description: Invalid user data (null or bad format)
        "409":
          description: Username or email already exists
        "422":
          description: Validation failed
  /User/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDTO"
            example:
              username: "admin"
              password: "Admin123!"
      responses:
        "200":
          description: Login successful, returns access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessfulLoginDTO"
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "422":
          description: User does not exist / invalid credentials / invalid data
  /User/accessToken:
    post:
      tags: [Auth]
      summary: Get new access token using refresh token cookie
      description: |
        Reads refresh token from cookie "RefreshToken",
        validates it and returns a new access token and refresh token.
      responses:
        "200":
          description: New access token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessfulLoginDTO"
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "401":
          description: Refresh token missing or invalid
        "422":
          description: User not found
  /User/logout:
    post:
      tags: [Auth]
      summary: Logout user
      description: Invalidates session and deletes refresh token cookie.
      responses:
        "200":
          description: Logout successful
        "401":
          description: Refresh token missing or invalid
  /User/me:
    get:
      tags: [Auth]
      summary: Get current user role and ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user role and id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleDTO"
              example:
                role: "Admin"
                id: "1"
        "404":
          description: User not found
  /User:
    get:
      tags: [Users]
      summary: Get all users (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAdminResponseDTO"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not Admin)
        "404":
          description: No users found
  /User/{id}:
    get:
      tags: [Users]
      summary: Get user by ID (self or Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (trying to access another user and not Admin)
        "404":
          description: User not found
  /User/block/{id}:
    post:
      tags: [Users]
      summary: Block user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User blocked successfully
          content:
            text/plain:
              schema:
                type: string
              example: "User 2 blocked successfully"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not Admin)
        "404":
          description: User not found
  /User/role:
    get:
      tags: [Roles]
      summary: Get all available roles (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not Admin)
        "404":
          description: No roles found
    post:
      tags: [Roles]
      summary: Change user role (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUserRoleDTO"
            example:
              id: 2
              newRole: "Admin"
      responses:
        "200":
          description: Role changed successfully
          content:
            text/plain:
              schema:
                type: string
              example: "User 2 role changed to Admin successfully"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not Admin)
        "404":
          description: User not found
        "500":
          description: Failed to remove or add user role

  # ---------------- COMMUNITIES ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegisterDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        name:
          type: string
        surname:
          type: string
    UserLoginDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    SuccessfulLoginDTO:
      type: object
      properties:
        accessToken:
          type: string
    UserRoleDTO:
      type: object
      properties:
        role:
          type: string
          description: "Admin or Member"
        id:
          type: string
    UserDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
        isBlocked:
          type: boolean
        role:
          type: string
    UserAdminResponseDTO:
      allOf:
        - $ref: "#/components/schemas/UserDto"
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        normalizedName:
          type: string
    ChangeUserRoleDTO:
      type: object
      properties:
        id:
          type: integer
        newRole:
          type: string
...